Revised manual analysis code
For data gathering on May 2. Initial plan is to save V over 1 s into the USB then process afterwards (so that uncertainty -- standard dev. -- can be calculated).

However, there are jumps in past data that may happen again, so maybe it is easier to have the oscilloscope display the stdev (if it can).